// uithub/routers/standard-url.d.ts
export type StandardURL = {
  omitFirstSegment?: boolean;
  primarySourceSegment: string;
  pluginId?: string;
  secondarySourceSegment?: string;
  basePath?: string;
  ext?: string;
  sourceType?: string;
  sourceUrl?: string;
  ogImageUrl?: string;
  title?: string;
  description?: string;
  rawUrlPrefix?: string;
};